refactor: extract window size variable into composable function
